// src/app/_core/domains/entities/meeyseo/enums/ms.structure.type.ts
export enum StructureType {
    Need = 'need',
    TypeOfHouse = 'typeOfHouse',
    TypeOfRealEstate = 'typeOfRealEstate',
    Project = 'project',
    Street = 'street',
    Ward = 'ward',
    District = 'district',
    City = 'city',
    Price = 'price',
    Area = 'area',
    Floor = 'floor',
    Bedroom = 'bedroom',
    Direction = 'direction',
    Facade = 'facade',
    Feature = 'feature',
    WideRoad = 'wideRoad',
    LegalPaper = 'legalPaper',
    HavingFE = 'havingFE',
    Utility = 'utility',
}